+++ AmbientCoreFramework/src/memory/ActionMemory.h
@@ -26,20 +26,20 @@ private:
      * @brief Unique identifier of the action that was performed
      * @invariant action_id >= 0
      */
-    int action_id;
+    int32_t action_id;
 
     /**
      * @brief Unique identifier of the entity that was used for the action
      * @invariant target_entity_id >= 0
      */
-    int target_entity_id;
+    int32_t target_entity_id;
 
 public:
     // =============================================================================
     // CONSTRUCTION
     // =============================================================================
 
-    explicit ActionMemory(int action_id, int target_entity_id, int last_used_time);
+    explicit ActionMemory(int32_t action_id, int32_t target_entity_id, int64_t last_used_time);
 
     // =============================================================================
     // MEMORY MATCHING
@@ -49,14 +49,14 @@ public:
     bool MatchesMemory(const BaseMemory& other) const override;
 
     [[nodiscard]]
-    bool MatchesMemory(int other_action_id, int other_target_entity_id) const;
+    bool MatchesMemory(int32_t other_action_id, int32_t other_target_entity_id) const;
 
     // =============================================================================
     // DATA ACCESS
     // =============================================================================
 
-    int GetActionId() const;
-    int GetTargetEntityId() const;
+    int32_t GetActionId() const;
+    int32_t GetTargetEntityId() const;
 
 };
